requirements.txt
# Core Data & ETL
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0

# MySQL Database
mysql-connector-python==8.2.0
pymysql==1.1.0
SQLAlchemy==2.0.23

# YouTube API
google-api-python-client==2.110.0

# Dashboard & Visualization
streamlit==1.29.0
plotly==5.18.0

# Machine Learning (forecasting)
scikit-learn==1.3.2

# AI/LLM Integration
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2

# Testing
pytest==7.4.3

# PDF Generation with charts
fpdf==1.7.2
reportlab==4.0.7
matplotlib==3.8.2

# Image Processing (for thumbnail A/B testing)
Pillow==10.1.0
